feat: extract the EditProfileModal
