+++ src/component/index.js
@@ -1,3 +1,4 @@
 export { AsideLeft } from "./AsideLeft";
 export { AsideRight } from "./AsideRight";
 export { Post } from "./Post";
+export { EditProfileModal } from "./EditProfileModal";
